# docs/api/backoffice-bff-openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/inventory": {
      "get": {
        "operationId": "InventoryController_getInventory",
        "summary": "在庫一覧取得",
        "parameters": [],
        "responses": {
          "200": {
            "description": "在庫一覧を返却"
          },
          "401": {
            "description": "認証エラー"
          }
        },
        "tags": [
          "inventory"
        ]
      }
    },
    "/api/bo/admin/inventory": {
      "get": {
        "operationId": "InventoryController_getInventory",
        "summary": "在庫一覧取得",
        "parameters": [],
        "responses": {
          "200": {
            "description": "在庫一覧を返却"
          },
          "401": {
            "description": "認証エラー"
          }
        },
        "tags": [
          "inventory"
        ]
      }
    },
...
    "/api/admin/items": {
      "get": {
        "operationId": "ProductsController_getProducts",
        "summary": "管理向け商品一覧取得",
        "parameters": [
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "商品一覧を返却"
          },
          "401": {
            "description": "認証エラー"
          }
        },
        "tags": [
          "products"
        ]
      },
      "post": {
        "operationId": "ProductsController_createProduct",
        "summary": "管理向け商品新規登録",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "作成後商品情報を返却"
          },
          "401": {
            "description": "認証エラー"
          }
        },
        "tags": [
          "products"
        ]
      }
    },
    "/api/bo/admin/items": {
      "get": {
        "operationId": "ProductsController_getProducts",
        "summary": "管理向け商品一覧取得",
        "parameters": [
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "商品一覧を返却"
          },
          "401": {
            "description": "認証エラー"
          }
        },
        "tags": [
          "products"
        ]
      },
      "post": {
        "operationId": "ProductsController_createProduct",
        "summary": "管理向け商品新規登録",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "作成後商品情報を返却"
          },
          "401": {
            "description": "認証エラー"
          }
        },
        "tags": [
          "products"
        ]
      }
    },
    "/api/admin/items/{id}": {
      "get": {
        "operationId": "ProductsController_getProductById",
        "summary": "管理向け商品詳細取得",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "商品詳細を返却"
          },
          "401": {
            "description": "認証エラー"
          }
        },
        "tags": [
          "products"
        ]
      },
      "put": {
        "operationId": "ProductsController_updateProduct",
        "summary": "管理向け商品更新",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "更新後商品情報を返却"
          },
          "401": {
            "description": "認証エラー"
          }
        },
        "tags": [
          "products"
        ]
      }
    },
    "/api/bo/admin/items/{id}": {
      "get": {
        "operationId": "ProductsController_getProductById",
        "summary": "管理向け商品詳細取得",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "商品詳細を返却"
          },
          "401": {
            "description": "認証エラー"
          }
        },
        "tags": [
          "products"
        ]
      },
      "put": {
        "operationId": "ProductsController_updateProduct",
        "summary": "管理向け商品更新",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "更新後商品情報を返却"
          },
          "401": {
            "description": "認証エラー"
          }
        },
        "tags": [
          "products"
        ]
      }
    },
    "/api/admin/item-categories": {
      "get": {
        "operationId": "ProductsController_getCategories",
        "summary": "管理向けカテゴリ一覧取得",
        "parameters": [],
        "responses": {
          "200": {
            "description": "カテゴリ一覧を返却"
          },
          "401": {
            "description": "認証エラー"
          }
        },
        "tags": [
          "products"
        ]
      },
      "post": {
        "operationId": "ProductsController_createCategory",
        "summary": "管理向けカテゴリ新規登録",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "作成後カテゴリ情報を返却"
          },
          "401": {
            "description": "認証エラー"
          }
        },
        "tags": [
          "products"
        ]
      }
    },
    "/api/bo/admin/item-categories": {
      "get": {
        "operationId": "ProductsController_getCategories",
        "summary": "管理向けカテゴリ一覧取得",
        "parameters": [],
        "responses": {
          "200": {
            "description": "カテゴリ一覧を返却"
          },
          "401": {
            "description": "認証エラー"
          }
        },
        "tags": [
          "products"
        ]
      },
      "post": {
        "operationId": "ProductsController_createCategory",
        "summary": "管理向けカテゴリ新規登録",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "作成後カテゴリ情報を返却"
          },
          "401": {
            "description": "認証エラー"
          }
        },
        "tags": [
          "products"
        ]
      }
    },
    "/api/admin/item-categories/{id}": {
      "put": {
        "operationId": "ProductsController_updateCategory",
        "summary": "管理向けカテゴリ更新",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "更新後カテゴリ情報を返却"
          },
          "401": {
            "description": "認証エラー"
          }
        },
        "tags": [
          "products"
        ]
      }
    },
    "/api/bo/admin/item-categories/{id}": {
      "put": {
        "operationId": "ProductsController_updateCategory",
        "summary": "管理向けカテゴリ更新",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "更新後カテゴリ情報を返却"
          },
          "401": {
            "description": "認証エラー"
          }
        },
        "tags": [
          "products"
        ]
      }
    }
  },
  "info": {
    "title": "BackOffice BFF API",
    "description": "管理向け BFF エンドポイント",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
...
}